Add confirmation popups for delete actions.

+++ app/src/main/java/com/elementary/tasks/birthdays/create/AddBirthdayActivity.kt
@@ -233,7 +233,9 @@ class AddBirthdayActivity : ThemedActivity<com.elementary.tasks.databinding.Acti
                 true
             }
             MENU_ITEM_DELETE -> {
-                deleteItem()
+                dialogues.askConfirmation(this, getString(R.string.delete)) {
+                    if (it) deleteItem()
+                }
                 true
             }
             else -> super.onOptionsItemSelected(item)
